toolkit/components/passwordmgr/test/mochitest/test_autocomplete_password_generation.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test autofill and autocomplete on autocomplete=new-password fields</title>
  <!-- This test assumes that telemetry events are not cleared after the `setup` task. -->
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script src="pwmgr_common.js"></script>
  <script src="../../../satchel/test/satchel_common.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css" />
</head>
<body>
Login Manager test: autofill with autocomplete=new-password fields

<script>
function initLogins() {
  Services.logins.removeAllUserFacingLogins();

  const {LoginManagerParent} = ChromeUtils.import("resource://gre/modules/LoginManagerParent.jsm");
  if (LoginManagerParent.getGeneratedPasswordsByPrincipalOrigin()) {
    LoginManagerParent.getGeneratedPasswordsByPrincipalOrigin().clear();
  }
}

runInParent(initLogins);

let origin = window.location.origin;
addLoginsInParent([origin, "https://autofill", null, "user1", "pass1"]);

let readyPromise = registerRunTests(3);
</script>
<p id="display"></p>

<!-- we presumably can't hide the content for this test. -->
<div id="content">

  <!-- form1 is the reference, sanity-check -->
  <form id="form1" action="https://autofill" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="p">
    <button type="submit">Submit</button>
  </form>

  <!-- form2 uses a new-password type=password field -->
  <form id="form2" action="https://autofill" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="password" autocomplete="new-password">
    <button type="submit">Submit</button>
  </form>

  <!-- form3 has enough signal to make new-password heuristics (Bug 1595244) detect
    a new password field, even though there is no autocomplete="new-password"
    attribute on the <input type="password"> element-->
  <form id="form3" action="https://autofill" onsubmit="return false;">
    <input type="text" name="username">
    <label>
      New password
      <input type="password" name="password">
    </label>
    <button type="submit">Submit</button>
  </form>
</div>

<pre id="test">
<script class="testbody" type="text/javascript">
const { ContentTaskUtils } = SpecialPowers.ChromeUtils.import(
  "resource://testing-common/ContentTaskUtils.jsm"
);
const { TestUtils } = SpecialPowers.ChromeUtils.import(
  "resource://testing-common/TestUtils.jsm"
);

let dateAndTimeFormatter = new SpecialPowers.Services.intl.DateTimeFormat(undefined, {
  dateStyle: "medium",
});

const TelemetryFilterPropsUsed = Object.freeze({
  category: "pwmgr",
  method: "autocomplete_field",
  object: "generatedpassword",
});

const TelemetryFilterPropsShown = Object.freeze({
  category: "pwmgr",
  method: "autocomplete_shown",
  object: "generatedpassword",
});

async function waitForTelemetryEventsCondition(cond, options = {},
                                               errorMsg = "waitForTelemetryEventsCondition timed out", maxTries = 200) {
  return TestUtils.waitForCondition(async () => {
    let events = await getTelemetryEvents(options);
    let result;
    try {
      result = cond(events);
      info("waitForTelemetryEventsCondition, result: " + result);
    } catch (e) {
      info("waitForTelemetryEventsCondition caught exception, got events: " + JSON.stringify(events));
      ok(false, `${e}\n${e.stack}`);
    }
    return result ? events : null;
  }, errorMsg, maxTries);
}

async function promiseACPopupClosed() {
  return SimpleTest.promiseWaitForCondition(async () => {
    let popupState = await getPopupState();
    return !popupState.open;
  }, "Wait for AC popup to be closed");
}

async function showACPopup(formNumber, expectedACLabels) {
  const autocompleteItems = await openAutocompletePopup();
  checkAutoCompleteResults(autocompleteItems, expectedACLabels,
                           window.location.host, "Check all rows are correct");
}

async function checkTelemetryEventsPWGenShown(expectedPWGenTelemetryEvents) {
  info(`showed generated password option, check there are now ${expectedPWGenTelemetryEvents} generatedpassword telemetry events`);
  await waitForTelemetryEventsCondition(events => {
    return events.length == expectedPWGenTelemetryEvents;
  }, { process: "parent", filterProps: TelemetryFilterPropsShown }, `Wait for there to be ${expectedPWGenTelemetryEvents} shown telemetry events`);
}

async function checkTelemetryEventsPWGenUsed(expectedPWGenTelemetryEvents) {
  info("filled generated password again, ensure we don't record another generatedpassword autocomplete telemetry event");
  let telemetryEvents;
  try {
    telemetryEvents = await waitForTelemetryEventsCondition(events => events.length == expectedPWGenTelemetryEvents + 1,
                                                            { process: "parent", filterProps: TelemetryFilterPropsUsed },
                                                            `Wait for there to be ${expectedPWGenTelemetryEvents + 1} used events`, 50);
  } catch (ex) {}
  ok(!telemetryEvents, `Expected to timeout waiting for there to be ${expectedPWGenTelemetryEvents + 1} events`);
}

add_setup(async () => {
  await SpecialPowers.pushPrefEnv({"set": [
    ["signon.generation.available", false],
    ["signon.generation.enabled", false],
  ]});
  ok(readyPromise, "check promise is available");
  await readyPromise;
  let useEvents = await getTelemetryEvents({ process: "parent", filterProps: TelemetryFilterPropsUsed, clear: true });
  is(useEvents.length, 0, "Expect 0 use events");
  let showEvents = await getTelemetryEvents({ process: "parent", filterProps: TelemetryFilterPropsShown, clear: true });
  is(showEvents.length, 0, "Expect 0 show events");
  let acEvents = await getTelemetryEvents({ process: "parent", filterProps: TelemetryFilterPropsAC, clear: true });
  is(acEvents.length, 0, "Expect 0 autocomplete events");

  // This can be removed once Bug 1623431 is resolved.
  document.getElementById("form3").reset();
});

add_task(async function test_autofillAutocompleteUsername_noGeneration() {
  // reference form was filled as expected?
  checkForm(1, "user1", "pass1");

  // 2nd form should not be filled
  checkForm(2, "", "");

  getFormElementByName(2, "uname").focus();
  await showACPopup(2, ["user1"]);

  let acEvents = await waitForTelemetryEventsCondition(events => {
    return events.length == 1;
  }, { process: "parent", filterProps: TelemetryFilterPropsAC, clear: true }, `Wait for there to be 1 autocomplete telemetry event`);
  checkACTelemetryEvent(acEvents[0], getFormElementByName(2, "uname"), {
    "hadPrevious": "0",
    "login": "1",
    "loginsFooter": "1"
  });

  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessedInSameProcess();
  checkForm(2, "user1", "pass1");

  recreateTree(document.getElementById("form2"));
});

add_task(async function test_autofillAutocompletePassword_noGeneration() {
  // 2nd form should not be filled
  checkForm(2, "", "");

  let pword = getFormElementByName(2, "password");
  pword.focus();
  await showACPopup(2, ["user1"]);
  let acEvents = await waitForTelemetryEventsCondition(events => {
    return events.length == 1;
  }, { process: "parent", filterProps: TelemetryFilterPropsAC, clear: true }, `Wait for there to be 1 autocomplete telemetry event`);
  checkACTelemetryEvent(acEvents[0], pword, {
    "hadPrevious": "0",
    "login": "1",
    "loginsFooter": "1"
  });

  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  // Can't use promiseFormsProcessedInSameProcess() when autocomplete fills the field directly.
  await SimpleTest.promiseWaitForCondition(() => pword.value == "pass1", "Check pw filled");
  checkForm(2, "", "pass1");

  // No autocomplete results should appear for non-empty pw fields.
  synthesizeKey("KEY_ArrowDown");
  await promiseNoUnexpectedPopupShown();

  recreateTree(document.getElementById("form2"));
});

// All tests below this are with generation prefs enabled.

add_task(async function test_autofillAutocompleteUsername_noGeneration2() {
  await SpecialPowers.pushPrefEnv({"set": [
    ["signon.generation.available", true],
    ["signon.generation.enabled", true],
  ]});

  // 2nd form should not be filled
  checkForm(2, "", "");

  getFormElementByName(2, "uname").focus();
  // No generation option on username fields.
  await showACPopup(2, ["user1"]);
  let acEvents = await waitForTelemetryEventsCondition(events => {
    return events.length == 1;
  }, { process: "parent", filterProps: TelemetryFilterPropsAC, clear: true }, `Wait for there to be 1 autocomplete telemetry event`);
  checkACTelemetryEvent(acEvents[0], getFormElementByName(2, "uname"), {
    "hadPrevious": "0",
    "login": "1",
    "loginsFooter": "1"
  });

  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessedInSameProcess();
  checkForm(2, "user1", "pass1");

  recreateTree(document.getElementById("form2"));
});

add_task(async function test_autofillAutocompletePassword_withGeneration() {
  const formNumbersToTest = [2, 3];
  // Bug 1616356 and Bug 1548878: Recorded once per origin
  let expectedPWGenTelemetryEvents = 0;
   // Bug 1619498: Recorded once every time the autocomplete popup is shown
  let expectedACShownTelemetryEvents = 0;
  for (let formNumber of formNumbersToTest) {
    // This form should not be filled
    checkForm(formNumber, "", "");

    let pword = getFormElementByName(formNumber, "password");
    pword.focus();

    await showACPopup(formNumber, [
      "user1",
      "Use a Securely Generated Password",
    ]);
    expectedPWGenTelemetryEvents++;
    expectedACShownTelemetryEvents++;

    await checkTelemetryEventsPWGenShown(expectedPWGenTelemetryEvents);
    let acEvents = await waitForTelemetryEventsCondition(events => {
      return events.length == expectedACShownTelemetryEvents;
    }, { process: "parent", filterProps: TelemetryFilterPropsAC }, `Wait for there to be ${expectedACShownTelemetryEvents} autocomplete telemetry event(s)`);
    checkACTelemetryEvent(acEvents[expectedACShownTelemetryEvents - 1], pword, {
      "generatedPasswo": "1",
      "hadPrevious": "0",
      "login": "1",
      "loginsFooter": "1"
    });

    synthesizeKey("KEY_ArrowDown");
    synthesizeKey("KEY_Enter");
    // Can't use promiseFormsProcessedInSameProcess() when autocomplete fills the field directly.
    await SimpleTest.promiseWaitForCondition(() => pword.value == "pass1", "Check pw filled");
    checkForm(formNumber, "", "pass1");

    // No autocomplete results should appear for non-empty pw fields.
    synthesizeKey("KEY_ArrowDown");
    await promiseNoUnexpectedPopupShown();

    info("Removing all logins to test auto-saving of generated passwords");
    await LoginManager.removeAllUserFacingLogins();

    while (pword.value) {
      synthesizeKey("KEY_Backspace");
    }
    LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "Blanked field");

    info("This time select the generated password");
    await showACPopup(formNumber, [
      "Use a Securely Generated Password",
    ]);
    expectedACShownTelemetryEvents++;

    await checkTelemetryEventsPWGenShown(expectedPWGenTelemetryEvents);
    acEvents = await waitForTelemetryEventsCondition(events => {
      return events.length == expectedACShownTelemetryEvents;
    }, { process: "parent", filterProps: TelemetryFilterPropsAC }, `Wait for there to be ${expectedACShownTelemetryEvents} autocomplete telemetry event(s)`);
    checkACTelemetryEvent(acEvents[expectedACShownTelemetryEvents - 1], pword, {
      "generatedPasswo": "1",
      "hadPrevious": "0",
      "loginsFooter": "1"
    });

    synthesizeKey("KEY_ArrowDown");
    let storageAddPromise = promiseStorageChanged(["addLogin"]);
    LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "Before first fill of generated pw");
    synthesizeKey("KEY_Enter");

    info("waiting for the password field to be filled with the generated password");
    await SimpleTest.promiseWaitForCondition(() => !!pword.value, "Check generated pw filled");
    LOGIN_FIELD_UTILS.checkPasswordMasked(pword, false, "After first fill of generated pw");
    info("Wait for generated password to be added to storage");
    await storageAddPromise;

    let logins = await LoginManager.getAllLogins();
    let timePasswordChanged = logins[logins.length - 1].timePasswordChanged;
    let time = dateAndTimeFormatter.format(new Date(timePasswordChanged));
    const LABEL_NO_USERNAME = "No username (" + time + ")";

    let generatedPW = pword.value;
    is(generatedPW.length, GENERATED_PASSWORD_LENGTH, "Check generated password length");
    ok(generatedPW.match(GENERATED_PASSWORD_REGEX), "Check generated password format");
    LOGIN_FIELD_UTILS.checkPasswordMasked(pword, false, "After fill");

    info("Check field is masked upon blurring");
    synthesizeKey("KEY_Tab"); // blur
    LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "After blur");
    synthesizeKey("KEY_Tab", { shiftKey: true }); // focus again
    LOGIN_FIELD_UTILS.checkPasswordMasked(pword, false, "After shift-tab to focus again");
    // Remove selection for OS where the whole value is selected upon focus.
    synthesizeKey("KEY_ArrowRight");

    while (pword.value) {
      LOGIN_FIELD_UTILS.checkPasswordMasked(pword, false, pword.value);
      synthesizeKey("KEY_Backspace");
    }
    LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "Blanked field");

    info("Blur the empty field to trigger a 'change' event");
    synthesizeKey("KEY_Tab"); // blur
    LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "Blur after blanking");
    synthesizeKey("KEY_Tab", { shiftKey: true }); // focus again
    LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "Focus again after blanking");

    info("Type a single character after blanking");
    synthesizeKey("@");

    info("Blur the single-character field to trigger a 'change' event");
    synthesizeKey("KEY_Tab"); // blur
    LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "Blur after backspacing");
    synthesizeKey("KEY_Tab", { shiftKey: true }); // focus again
    LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "Focus again after backspacing");
    synthesizeKey("KEY_Backspace"); // Blank the field again

    await showACPopup(formNumber, [
      LABEL_NO_USERNAME,
      "Use a Securely Generated Password",
    ]);
    expectedACShownTelemetryEvents++;

    synthesizeKey("KEY_ArrowDown");
    synthesizeKey("KEY_ArrowDown");
    synthesizeKey("KEY_Enter");
    await SimpleTest.promiseWaitForCondition(() => !!pword.value, "Check generated pw filled");
    // Same generated password should be used, even despite the 'change' to @ earlier.
    checkForm(formNumber, "", generatedPW);
    LOGIN_FIELD_UTILS.checkPasswordMasked(pword, false, "Second fill of the generated pw");

    await checkTelemetryEventsPWGenUsed(expectedPWGenTelemetryEvents);

    logins = await LoginManager.getAllLogins();
    is(logins.length, 1, "Still 1 login after filling the generated password a 2nd time");
    is(logins[0].timePasswordChanged, timePasswordChanged, "Saved login wasn't changed");
    is(logins[0].password, generatedPW, "Password is the same");

    info("filling the saved login to ensure the field is masked again");

    while (pword.value) {
      LOGIN_FIELD_UTILS.checkPasswordMasked(pword, false, pword.value);
      synthesizeKey("KEY_Backspace");
    }
    LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "Blanked field again");

    info("Blur the field to trigger a 'change' event again");
    synthesizeKey("KEY_Tab"); // blur
    LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "Blur after blanking again");
    synthesizeKey("KEY_Tab", { shiftKey: true }); // focus again
    LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "Focus again after blanking again");
    // Remove selection for OS where the whole value is selected upon focus.
    synthesizeKey("KEY_ArrowRight");

    await showACPopup(formNumber, [
      LABEL_NO_USERNAME,
      "Use a Securely Generated Password",
    ]);
    expectedACShownTelemetryEvents++;

    synthesizeKey("KEY_ArrowDown");
    synthesizeKey("KEY_Enter");
    await SimpleTest.promiseWaitForCondition(() => !!pword.value, "Check saved generated pw filled");
    // Same generated password should be used but from storage
    checkForm(formNumber, "", generatedPW);
    // Passwords from storage should always be masked.
    LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "after fill from storage");
    synthesizeKey("KEY_Tab"); // blur
    LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "after blur");
    synthesizeKey("KEY_Tab", { shiftKey: true }); // focus
    LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "after shift-tab to focus again");

    info("reset initial login state with login1");
    runInParent(initLogins);
    addLoginsInParent([origin, "https://autofill", null, "user1", "pass1"]);

    info("invalidate the autocomplete cache after updating storage above");
    synthesizeKey("KEY_Backspace");

    recreateTree(document.getElementById(`form${formNumber}`));
  }
});

add_task(async function test_autofillAutocompletePassword_saveLoginDisabled() {
  // form should not be filled
  checkForm(2, "", "");
  let formOrigin = new URL(document.documentURI).origin;
  is(formOrigin, origin, "Expected form origin");

  await LoginManager.setLoginSavingEnabled(origin, false);

  let pword = getFormElementByName(2, "password");
  pword.focus();
  // when login-saving is disabled for an origin, we expect no generated password row here
  await showACPopup(2, ["user1"]);

  // close any open menu
  synthesizeKey("KEY_Escape");
  await promiseACPopupClosed();

  recreateTree(document.getElementById("form2"));
  await LoginManager.setLoginSavingEnabled(origin, true);
});

add_task(async function test_deleteAndReselectGeneratedPassword() {
  info("Removing all logins to test auto-saving of generated passwords");
  await LoginManager.removeAllUserFacingLogins();

  // form should not be filled
  checkForm(2, "", "");
  let pword = getFormElementByName(2, "password");
  let uname = getFormElementByName(2, "uname");

  async function showAndSelectACPopupItem(index) {
    pword.focus();
    if (pword.value) {
      pword.select();
      synthesizeKey("KEY_Backspace");
    }
    const autocompleteItems = await openAutocompletePopup();
    if (index < 0) {
      index = autocompleteItems.length + index;
    }
    for (let i=0; i<=index; i++) {
      synthesizeKey("KEY_ArrowDown");
    }
    await TestUtils.waitForTick();
    return autocompleteItems[index];
  }

  let storagePromise, menuLabel, itemIndex, savedLogins;

  // fill the password field with the generated password via auto-complete menu
  storagePromise = promiseStorageChanged(["addLogin"]);
  // select last-but-2 item - the one before the footer
  menuLabel = await showAndSelectACPopupItem(-2);
  is(menuLabel, "Use a Securely Generated Password", "Check item label");
  synthesizeKey("KEY_Enter");
  info("waiting for the password field to be filled with the generated password");
  await SimpleTest.promiseWaitForCondition(() => !!pword.value, "Check generated pw filled");
  info("Wait for generated password to be added to storage");
  await storagePromise;

  uname.focus();
  await TestUtils.waitForTick();

  is(pword.value.length, LoginTestUtils.generation.LENGTH, "Check password looks generated");
  const GENERATED_PASSWORD = pword.value;

  savedLogins = await LoginManager.getAllLogins();
  is(savedLogins.length, 1, "Check saved logins count");

  info("clear the password field and delete the saved login using the AC menu")
  storagePromise = promiseStorageChanged(["removeLogin"]);

  itemIndex = 0;
  menuLabel = await showAndSelectACPopupItem(itemIndex);
  ok(menuLabel.includes("No username"), "Check first item is the auto-saved login");
  // Send delete to remove the auto-saved login from storage
  // On OS X, shift-backspace and shift-delete work, just delete does not.
  // On Win/Linux, shift-backspace does not work, delete and shift-delete do.
  synthesizeKey("KEY_Delete", {shiftKey: true});
  await storagePromise;

  uname.focus();
  await TestUtils.waitForTick();

  savedLogins = await LoginManager.getAllLogins();
  is(savedLogins.length, 0, "Check saved logins count");

  info("Re-fill with the generated password");
  // select last-but-2 item - the one before the footer
  menuLabel = await showAndSelectACPopupItem(-2);
  is(menuLabel, "Use a Securely Generated Password", "Check item label");
  synthesizeKey("KEY_Enter");
  info("waiting for the password field to be filled with the generated password");
  await SimpleTest.promiseWaitForCondition(() => !!pword.value, "Check generated pw filled");

  uname.focus();
  await TestUtils.waitForTick();
  is(pword.value, GENERATED_PASSWORD, "Generated password has not changed");

  recreateTree(document.getElementById("form2"));
});

add_task(async function test_passwordGenerationShownTelemetry() {
  // Should only be recorded once per principal origin per session, but the cache is cleared each time ``initLogins`` is called.
  await waitForTelemetryEventsCondition(events => {
    return events.length == 3;
  }, { process: "parent", filterProps: TelemetryFilterPropsShown }, "Expect 3 shown telemetry events");
});
</script>
</pre>
</body>
</html>
